Moving some exports around
