+++ packages/keystatic/src/app/useItemData.ts
@@ -18,7 +18,7 @@ import {
   KEYSTATIC_CLOUD_HEADERS,
   MaybePromise,
 } from './utils';
-import { toFormattedFormDataError } from '../form/errors';
+import { toFormattedFormDataError } from '../form/error-formatting';
 import { serializeRepoConfig } from './repo-config';
 
 function parseEntry(args: UseItemDataArgs, files: Map<string, Uint8Array>) {
